requirements.txt
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
pycodestyle==2.10.0
pydantic==1.10.7
PyJWT==2.7.0
python-dotenv==1.0.0
redis==4.5.5
requests==2.30.0
rq==1.15.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
